+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <service
-            android:name=".presenter.view.FcmService"
+            android:name=".presenter.service.FcmService"
             android:enabled="true"
             android:exported="false">
             <intent-filter>
