+++ app/src/main/java/campus/tech/kakao/map/presenter/service/FcmService.kt
@@ -7,7 +7,6 @@ import android.content.Context
 import android.content.Intent
 import androidx.core.app.NotificationCompat
 import campus.tech.kakao.map.R
-import campus.tech.kakao.map.base.BaseNotificationService
 import campus.tech.kakao.map.presenter.view.SplashActivity
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
@@ -17,7 +16,53 @@ class FcmService : FirebaseMessagingService() {
     override fun onMessageReceived(message: RemoteMessage) {
         super.onMessageReceived(message)
 
-        val notificationService = BaseNotificationService(this)
-        notificationService.notifyFromFCM(message.toString())
+        var notificationManager: NotificationManager = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+
+        notificationManager.createNotificationChannel(createNotificationChannel())
+        val intent = Intent(this, SplashActivity::class.java).apply {
+            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
+        }
+        val pendingIntent = PendingIntent.getActivity(
+            this,
+            0,
+            intent,
+            PendingIntent.FLAG_IMMUTABLE
+        )
+        val builder = NotificationCompat.Builder(
+            this,
+            CHANNEL_ID
+        )
+            .setSmallIcon(R.drawable.ic_launcher_foreground)
+            .setContentTitle("[중요] 포그라운드 알림")
+            .setContentText("앱이 실행 중입니다.")
+            .setPriority(NotificationCompat.PRIORITY_DEFAULT)
+            .setContentIntent(pendingIntent)
+            .setStyle(
+                NotificationCompat.BigTextStyle()
+                    .bigText("앱이 실행 중일 때는 포그라운드 알림이 발생합니다.")
+            )
+            .setAutoCancel(true)
+
+        notificationManager.notify(NOTIFICATION_ID, builder.build())
+    }
+
+    private fun createNotificationChannel() : NotificationChannel {
+        return NotificationChannel(
+            CHANNEL_ID,
+            CHANNEL_NAME,
+            NotificationManager.IMPORTANCE_DEFAULT
+        ).apply {
+            description = CHANNEL_DESCRIPTION
+        }
+    }
+
+
+
+    companion object {
+        private const val NOTIFICATION_ID = 222222
+        private const val CHANNEL_ID = "main_default_channel"
+        private const val CHANNEL_NAME = "main channelName"
+        private const val CHANNEL_DESCRIPTION = "main channelDescription"
+
     }
 }
